backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
openai
python-multipart
pydantic
python-docx
openpyxl
